Split player class and playback start
