+++ spotipy/client/player/__init__.py
@@ -1,0 +1,6 @@
+from spotipy.client.player.modify import SpotifyPlayerModify
+from spotipy.client.player.view import SpotifyPlayerView
+
+
+class SpotifyPlayer(SpotifyPlayerModify, SpotifyPlayerView):
+    pass
